core/requirements.txt
flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
flask-login==0.6.3
WTForms==3.1.2
email-validator==2.2.0